Explode settings into multiple files. This seems to help TSLint find all the type information it wants so it stops complaining elsewhere.
