+++ packages/app/src/context/settings/index.ts
@@ -1,0 +1,2 @@
+export { useSettings } from './context';
+export { Provider } from './provider';
