+++ packages/app/src/context/settings/types.ts
@@ -1,0 +1,18 @@
+export type BoolOrNull = boolean | null;
+
+export type State = {
+  dark: BoolOrNull;
+};
+
+export type ActionType = 'load' | 'setdark';
+
+export type Action = {
+  type: ActionType;
+  dark?: BoolOrNull;
+  state?: State;
+};
+
+export type Context = {
+  state: State;
+  dispatch: React.Dispatch<Action>;
+};
